+++ app/code/autoStore/Sample/etc/module.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="autoStore_Sample" setup_version="1.0.0">
+        <sequence>
+            <module name="Magento_Cms" />
+        </sequence>
+    </module>
+</config>
