+++ app/code/autoStore/Sample/registration.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+
+autoStore module.
+@author Vladislav Romaniuk vlad96r@protonmial.com
+@copyright 2020 Vlromaniuk
+ */
+
+
+use \Magento\Framework\Component\ComponentRegistrar;
+
+ComponentRegistrar::register(ComponentRegistrar::MODULE, 'autoStore_Sample', __DIR__);
