+++ app/code/autoStore/Sample/view/frontend/layout/cms_index_index.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="2columns-left" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceContainer name="content">
+                <block class="Magento\Framework\View\Element\Template"
+                       name="custom.module"
+                       template="autoStore_Sample::custom_1.phtml"/>
+        </referenceContainer>
+    </body>
+</page>
